notification/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-phonenumber-field==7.0.2
django-tinymce==3.5.0
idna==3.4
pendulum==2.1.2
phonenumbers==8.13.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pytzdata==2020.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
twilio==7.16.3
urllib3==1.26.14
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-phonenumber-field==7.0.2
django-tinymce==3.5.0
idna==3.4
pendulum==2.1.2
phonenumbers==8.13.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pytzdata==2020.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
twilio==7.16.3
urllib3==1.26.14
